requirements.txt
Django==2.2.4
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-richtextfield==1.4.0
Pillow==6.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2019.2
sqlparse==0.3.0
